Change evaluation to int16_t from int32_t

+++ Captain/engine.cpp
@@ -147,14 +147,14 @@ namespace engine
 		movegen::Movelist<218> moves;
 		movegen::genMoves(b, moves);
 		
-		std::array<std::pair<board::Move, std::int32_t>, 218> rootMoves;
+		std::array<std::pair<board::Move, std::int16_t>, 218> rootMoves;
 		for (std::size_t i = 0; i != moves.size(); ++i)
 		{
 			rootMoves[i].first = moves[i];
 			rootMoves[i].second = negInf;
 		}
 
-		std::int32_t worstCase = negInf;
+		std::int16_t worstCase = negInf;
 
 		board::QBB bcopy = b;
 		for (unsigned int k = 1; k <= posInf; ++k)
@@ -162,7 +162,7 @@ namespace engine
 			sync_cout << "info string iterative deepening " << k << sync_endl;
 			currIDdepth = k;
 			worstCase = negInf;
-			std::int32_t score = negInf;
+			std::int16_t score = negInf;
 
 			for (std::size_t i = 0; i != moves.size(); ++i)
 			{
@@ -198,9 +198,9 @@ namespace engine
 		sync_cout << "bestmove " << move2uciFormat(rootMoves[0].first) << sync_endl;
 	}
 
-	std::int32_t Engine::quiesceSearch(const board::QBB& b, std::int32_t alpha, std::int32_t beta, int depth)
+	std::int16_t Engine::quiesceSearch(const board::QBB& b, std::int16_t alpha, std::int16_t beta, int depth)
 	{
-		const std::int32_t checkpos = eval::evaluate(b);
+		const std::int16_t checkpos = eval::evaluate(b);
 		if (b.get50() == 50)
 			return 0;
 		if (checkpos > beta)
@@ -243,7 +243,7 @@ namespace engine
 			}
 		}
 
-		std::int32_t currEval = checkpos;
+		std::int16_t currEval = checkpos;
 
 		board::QBB bcopy = b;
 		for (std::size_t i = 0; i != ml.size(); ++i)
@@ -263,7 +263,7 @@ namespace engine
 		return currEval;
 	}
 
-	std::int32_t Engine::alphaBetaSearch(const board::QBB& b, std::int32_t alpha, std::int32_t beta, int depth, bool nullBranch)
+	std::int16_t Engine::alphaBetaSearch(const board::QBB& b, std::int16_t alpha, std::int16_t beta, int depth, bool nullBranch)
 	{
 		const auto oldAlpha = alpha;
 		if (shouldStop())
@@ -314,7 +314,7 @@ namespace engine
 				return 0;
 		}
 
-		std::int32_t currEval = negInf;
+		std::int16_t currEval = negInf;
 
 		
 		board::QBB bcopy = b;
