+++ Captain/transpositiontable.hpp
@@ -32,7 +32,7 @@ namespace TTable
 	struct Entry
 	{
 		std::uint64_t key = 0;
-		std::int32_t eval = 0;
+		std::int16_t eval = 0;
 		int depth = 0;
 		char nodeType = 0; // 0 = PV (exact score), 1 = all (upper bound), 2 = cut (lower bound)
 	};
@@ -58,7 +58,7 @@ namespace TTable
 		
 		void clear();
 
-		void store(std::uint64_t hash, int depth, std::int32_t eval, char nodetype);
+		void store(std::uint64_t hash, int depth, std::int16_t eval, char nodetype);
 
 		Entry& operator[](std::uint64_t hash) noexcept;
